add PIDTracerV3

+++ command/ColorPIDTracerV2.h
@@ -28,7 +28,7 @@ private:
     float r = 0;
     rgb_raw_t targetRGB;
     float beforeP = 0;
-    float integral = 0;
+    double integral = 0;
     PIDTracerMode traceMode;
     TraceColor traceColor;
